Remove Utf8String and rename to UTF8Buffer
